Add update/delete api in buffer, rename middleware
