+++ ding/worker/buffer/buffer.py
@@ -1,5 +1,6 @@
 from typing import Any, Callable, List
 from ding.worker.buffer.storage import Storage
+import uuid
 import copy
 
 
@@ -9,24 +10,24 @@ def apply_middleware(func_name: str):
 
         def handler(buffer, *args, **kwargs):
             """
-            The real processing starts here, we apply the middlewares one by one,
+            The real processing starts here, we apply the middleware one by one,
             each middleware will receive a `next` function, which is an executor of next
             middleware. You can change the input arguments to the `next` middleware, and you
             also can get the return value from the next middleware, so you have the
             maximum freedom to choose at what stage to implement your method.
             """
 
-            def wrap_handler(middlewares, *args, **kwargs):
-                if len(middlewares) == 0:
+            def wrap_handler(middleware, *args, **kwargs):
+                if len(middleware) == 0:
                     return base_func(buffer, *args, **kwargs)
 
                 def next(*args, **kwargs):
-                    return wrap_handler(middlewares[1:], *args, **kwargs)
+                    return wrap_handler(middleware[1:], *args, **kwargs)
 
-                func = middlewares[0]
+                func = middleware[0]
                 return func(func_name, next, *args, **kwargs)
 
-            return wrap_handler(buffer.middlewares, *args, **kwargs)
+            return wrap_handler(buffer.middleware, *args, **kwargs)
 
         return handler
 
@@ -43,7 +44,7 @@ class Buffer:
             - storage (:obj:`Storage`): The storage instance.
         """
         self.storage = storage
-        self.middlewares = []
+        self.middleware = []
 
     @apply_middleware("push")
     def push(self, data: Any) -> None:
@@ -59,7 +60,7 @@ class Buffer:
     def sample(self, size: int, replace: bool = False, range: slice = None) -> List[Any]:
         """
         Overview:
-            Sample data with length ``size``, this function may be wrapped by middlewares.
+            Sample data with length ``size``, this function may be wrapped by middleware.
         Arguments:
             - size (:obj:`int`): The number of the data that will be sampled.
         Returns:
@@ -75,10 +76,16 @@ class Buffer:
         """
         self.storage.clear()
 
+    def update(self, index: str, data: Any) -> None:
+        pass
+
+    def delete(self, index: str) -> None:
+        pass
+
     def use(self, func: Callable) -> "Buffer":
         r"""
         Overview:
-            Use algorithm middlewares to modify the behavior of the buffer.
+            Use algorithm middleware to modify the behavior of the buffer.
             Every middleware should be a callable function, it will receive three argument parts, including:
             1. The buffer instance, you can use this instance to visit every thing of the buffer,
                including the storage.
@@ -88,14 +95,14 @@ class Buffer:
 
             Each middleware handler should return two parts of the value, including:
             1. The first value is `done` (True or False), if done==True, the middleware chain will stop immediately,
-               no more middlewares will be executed during this execution
+               no more middleware will be executed during this execution
             2. The remaining values, will be passed to the next middleware or the default function in the buffer.
         Arguments:
             - func (:obj:`Callable`): The middleware handler
         Returns:
             - buffer (:obj:`Buffer`): The instance self
         """
-        self.middlewares.append(func)
+        self.middleware.append(func)
         return self
 
     def view(self) -> "Buffer":
@@ -107,5 +114,5 @@ class Buffer:
             - buffer (:obj:`Buffer`): The instance self
         """
         buffer = Buffer(self.storage)
-        buffer.middlewares = copy.deepcopy(self.middlewares)
+        buffer.middleware = copy.deepcopy(self.middleware)
         return buffer
