+++ ding/worker/buffer/memory_storage.py
@@ -1,6 +1,8 @@
-from typing import Any, List
+from typing import Any, List, Optional
 from collections import deque
 from operator import itemgetter
+
+from numpy import delete
 from ding.worker.buffer import Storage
 import itertools
 import random
@@ -17,7 +19,7 @@ class MemoryStorage(Storage):
     def get(self, indices: List[int]) -> List[Any]:
         return itemgetter(*indices)(self.storage)
 
-    def sample(self, size: int, replace: bool = False, range: slice = None) -> List[Any]:
+    def sample(self, size: int, replace: bool = False, range: Optional[slice] = None) -> List[Any]:
         storage = self.storage
         if range:
             storage = list(itertools.islice(self.storage, range.start, range.stop, range.step))
@@ -25,6 +27,12 @@ class MemoryStorage(Storage):
             return random.choices(storage, k=size)
         else:
             return random.sample(storage, k=size)
+
+    def update(self, index: str, data: Any, extra: Optional[Any] = None) -> None:
+        pass
+
+    def delete(self, index: str) -> None:
+        pass
 
     def count(self) -> int:
         return len(self.storage)
