+++ ding/worker/buffer/middleware/clone_object.py
