+++ ding/worker/buffer/storage.py
@@ -1,11 +1,11 @@
 from abc import abstractmethod
-from typing import Any, List
+from typing import Any, List, Optional
 
 
 class Storage:
 
     @abstractmethod
-    def append(self, data: Any) -> None:
+    def append(self, data: Any, index: Optional[str] = None, extra: Optional[dict] = None) -> None:
         raise NotImplementedError
 
     @abstractmethod
@@ -13,7 +13,15 @@ class Storage:
         raise NotImplementedError
 
     @abstractmethod
-    def sample(self, size: int, replace: bool = False, range: slice = None) -> List[Any]:
+    def sample(self, size: int, replace: bool = False, range: Optional[slice] = None) -> List[Any]:
+        raise NotImplementedError
+
+    @abstractmethod
+    def update(self, index: str, data: Any, extra: Optional[Any] = None) -> None:
+        raise NotImplementedError
+
+    @abstractmethod
+    def delete(self, index: str) -> None:
         raise NotImplementedError
 
     @abstractmethod
